Clean up tests and mocks
