+++ app/client/components/DownloadList/DownloadList.test.js
@@ -4,40 +4,20 @@ import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import 'isomorphic-fetch';
 import DownloadList from '../DownloadList';
+import {
+  downloadListMock,
+  paginatedDownloadListMock,
+  filteredDownloadListMock,
+} from 'Utils/mocks';
 
 test('should fetch and render download list', async () => {
-  const fakeDownloadList = {
-    downloads: [
-      {
-        id: 0,
-        name: 'Download One',
-        status: 'In Progress',
-        percent: 50.0,
-        size: '30.6MiB',
-        speed: '100KiB/s',
-        timeRemaining: '00:38',
-      },
-      {
-        id: 1,
-        name: 'Download Two',
-        status: 'In Progress',
-        percent: 100.0,
-        size: '52.8MiB',
-        speed: '0KiB/s',
-        timeRemaining: '00:00',
-      },
-    ],
-    totalPages: 1,
-  };
+  // GIVEN
+  downloadListMock();
 
-  jest.spyOn(global, 'fetch').mockImplementation(() => {
-    return Promise.resolve({
-      json: () => Promise.resolve(fakeDownloadList),
-    });
-  });
-
+  // WHEN
   const { findByText } = render(<DownloadList />);
 
+  // THEN
   const firstDownloadText = await findByText('Download One');
   expect(firstDownloadText).toBeInTheDocument();
 
@@ -70,179 +50,61 @@ test('should fetch and render download list', async () => {
 });
 
 test('should load specific page when pagination link clicked on', async () => {
-  const pageOneFakeDownloadList = {
-    downloads: [
-      {
-        id: 0,
-        name: 'Download One',
-        status: 'In Progress',
-        percent: 50.0,
-        size: '30.6MiB',
-        speed: '100KiB/s',
-        timeRemaining: '00:38',
-      },
-      {
-        id: 1,
-        name: 'Download Two',
-        status: 'In Progress',
-        percent: 100.0,
-        size: '52.8MiB',
-        speed: '0KiB/s',
-        timeRemaining: '00:00',
-      },
-    ],
-    totalPages: 2,
-  };
-  const pageTwoFakeDownloadList = {
-    downloads: [
-      {
-        id: 2,
-        name: 'Download Three',
-        status: 'In Progress',
-        percent: 50.0,
-        size: '30.6MiB',
-        speed: '100KiB/s',
-        timeRemaining: '00:38',
-      },
-      {
-        id: 3,
-        name: 'Download Four',
-        status: 'In Progress',
-        percent: 100.0,
-        size: '52.8MiB',
-        speed: '0KiB/s',
-        timeRemaining: '00:00',
-      },
-    ],
-    totalPages: 2,
-  };
+  // GIVEN
+  paginatedDownloadListMock();
 
-  jest.spyOn(global, 'fetch').mockImplementation(url => {
-    if (url.includes('2')) {
-      return Promise.resolve({
-        json: () => Promise.resolve(pageTwoFakeDownloadList),
-      });
-    }
-
-    return Promise.resolve({
-      json: () => Promise.resolve(pageOneFakeDownloadList),
-    });
-  });
-
+  // WHEN
   const { findByText } = render(<DownloadList />);
 
+  // THEN
   const firstPageDownloadText = await findByText('Download One');
   expect(firstPageDownloadText).toBeInTheDocument();
 
+  // WHEN
   const pageTwoButton = await findByText('2');
   fireEvent.click(pageTwoButton);
 
+  // THEN
   const secondPageDownloadText = await findByText('Download Three');
   expect(secondPageDownloadText).toBeInTheDocument();
 });
 
 test('should load specific status when status filter clicked on', async () => {
-  const pageOneInProgressDownloadList = {
-    downloads: [
-      {
-        id: 1,
-        name: 'Download One',
-        status: 'In Progress',
-        percent: 50.0,
-        size: '30.6MiB',
-        speed: '100KiB/s',
-        timeRemaining: '00:38',
-      },
-    ],
-    totalPages: 2,
-  };
-  const pageOneCompletedDownloadList = {
-    downloads: [
-      {
-        id: 3,
-        name: 'Download Three',
-        status: 'Completed',
-        percent: 100.0,
-        size: '30.6MiB',
-        speed: '100KiB/s',
-        timeRemaining: '00:38',
-      },
-    ],
-    totalPages: 2,
-  };
-  const pageTwoInProgressDownloadList = {
-    downloads: [
-      {
-        id: 2,
-        name: 'Download Two',
-        status: 'In Progress',
-        percent: 50.0,
-        size: '30.6MiB',
-        speed: '100KiB/s',
-        timeRemaining: '00:38',
-      },
-    ],
-    totalPages: 2,
-  };
-  const pageTwoCompletedDownloadList = {
-    downloads: [
-      {
-        id: 4,
-        name: 'Download Four',
-        status: 'Completed',
-        percent: 100.0,
-        size: '30.6MiB',
-        speed: '100KiB/s',
-        timeRemaining: '00:38',
-      },
-    ],
-    totalPages: 2,
-  };
+  // GIVEN
+  filteredDownloadListMock();
 
-  jest.spyOn(global, 'fetch').mockImplementation(url => {
-    if (url.includes('Completed,Error') && url.includes('1')) {
-      return Promise.resolve({
-        json: () => Promise.resolve(pageOneCompletedDownloadList),
-      });
-    } else if (url.includes('Completed,Error') && url.includes('2')) {
-      return Promise.resolve({
-        json: () => Promise.resolve(pageTwoCompletedDownloadList),
-      });
-    } else if (url.includes('In Progress') && url.includes('2')) {
-      return Promise.resolve({
-        json: () => Promise.resolve(pageTwoInProgressDownloadList),
-      });
-    }
-
-    return Promise.resolve({
-      json: () => Promise.resolve(pageOneInProgressDownloadList),
-    });
-  });
-
+  // WHEN
   const { findByText } = render(<DownloadList />);
 
+  // THEN
   const firstPageInProgressDownloadText = await findByText('IN PROGRESS');
   expect(firstPageInProgressDownloadText).toBeInTheDocument();
   const firstPageInProgressDownloadNameText = await findByText('Download One');
   expect(firstPageInProgressDownloadNameText).toBeInTheDocument();
 
+  // WHEN
   const completedButton = await findByText('Completed Error');
   fireEvent.click(completedButton);
 
+  // THEN
   const firstPageCompletedDownloadText = await findByText('COMPLETED');
   expect(firstPageCompletedDownloadText).toBeInTheDocument();
   const firstPageCompletedDownloadNameText = await findByText('Download Three');
   expect(firstPageCompletedDownloadNameText).toBeInTheDocument();
 
+  // WHEN
   const secondPageButton = await findByText('2');
   fireEvent.click(secondPageButton);
 
+  // THEN
   const secondPageCompletedDownloadNameText = await findByText('Download Four');
   expect(secondPageCompletedDownloadNameText).toBeInTheDocument();
 
+  // WHEN
   const inProgressButton = await findByText('In Progress');
   fireEvent.click(inProgressButton);
 
+  // THEN
   const secondPageInProgressDownloadText = await findByText('IN PROGRESS');
   expect(secondPageInProgressDownloadText).toBeInTheDocument();
   const secondPageInProgressDownloadNameText = await findByText('Download One');
