+++ app/client/components/DownloadModal/DownloadModal.test.js
@@ -6,21 +6,27 @@ import 'isomorphic-fetch';
 import DownloadModal from '../DownloadModal';
 
 test('should open DownloadModal on button click', () => {
+  // WHEN
   const { getByText } = render(<DownloadModal />);
   fireEvent.click(getByText('New Download...'));
+
+  // THEN
   expect(getByText('URL')).toBeInTheDocument();
   expect(getByText('Directory')).toBeInTheDocument();
 });
 
 test('should POST and close modal on Download click', () => {
+  // GIVEN
   jest.spyOn(global, 'fetch').mockImplementation(() => {
     return Promise.resolve({});
   });
 
+  // THEN
   const { getByText, queryByText } = render(<DownloadModal />);
   fireEvent.click(getByText('New Download...'));
   fireEvent.click(getByText('Download'));
 
+  // THEN
   expect(queryByText('URL')).not.toBeInTheDocument();
   expect(queryByText('Directory')).not.toBeInTheDocument();
 
@@ -28,9 +34,12 @@ test('should POST and close modal on Download click', () => {
 });
 
 test('should close on Cancel click', () => {
+  // WHEN
   const { getByText, queryByText } = render(<DownloadModal />);
   fireEvent.click(getByText('New Download...'));
   fireEvent.click(getByText('Cancel'));
+
+  // THEN
   expect(queryByText('URL')).not.toBeInTheDocument();
   expect(queryByText('Directory')).not.toBeInTheDocument();
 });
